backend/requirements.txt
# FastAPI 및 웹 서버
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Supabase 클라이언트
supabase==2.3.4

# 환경 변수 및 설정
python-dotenv==1.0.0
pydantic-settings==2.1.0

# 인증 및 보안
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.26.0

# 유틸리티
python-dateutil==2.8.2
